Define booleans in new macros module
A new macros module has been made to gather all common functions and
macros for all other modules. The first macros to be defined are the
boolean TRUE|FALSE.

As TRUE can be any value but FALSE is always 0, TRUE has been defined
as the contrary of FALSE and not to 1. This allows booleans to be used
even if the returned value is not 0 or 1.