Implement collision detection
In order to be able to correctly do collision detection, a collision box
had to be created. If not, the character would be blocked most of the
time in hallways of its size. Moreover, this collision box allows a
type of fake 3d as it is place on the character feet, thus most of it's
body can go through the collision tile.