Switch from SDL1 to SDL2
Since start the GUI programmation was done with the SDL1 framework. This
commit upgrade to SDL2. All functions that needed to be changed were
done, particularly the switch from Surface to Renderer. Some other
portions of code were also adapted to the new environment, esp. the
tiles size in main (no more rotozoom), or the pause red rectangle that
needed to be created first.