Fix tiles not rendering properly
Tiles were not rendered as expected, as they were offsetted and were not
making a continuous surface with each other. Tiles had to be adapted to
the isometric set up, ie. TILES_HEIGHT needed to be half of TILES_WIDTH
(in fact TILES_WIDTH was doubled), and the assets were modified
properly. Moreover, the level 1 tiles (FOREGROUND), needed to be
offsetted upward to reflect the fake Z axis.