Set own macros for PF test unit
As TILES_WIDTH and TILES_HEIGHT have changed to be adapted to isometric
projection, the PF test unit had to be adapted. Instead of making an
isometric projection too in the unit, which is not the point of the
test, own width and height were defined, and the whole unit adapted.
STARTX and Y were also defined for consistency.