Fix lag in mouse hovering
If a lot of event was to handle by the event loop, the mouse hover
could show some lag and the cursor was chasing the real mouse position.
Two things needed to be fixed : the re-drawing of screen was done
constantely and not at FRAMERATE, putting a heavy burden on CPU, and the
PollEvent was called even if no event was here, re-sending previous
events in queue.