Merge all objects structures into one
All objects should be managed by the same structure, for convenience.
This allows to pass around only one array and pick the right object by
its hex ID, instead of multiplying arrays for different types.
This commit creates the Asset structure, describing each objects (tiles,
characters, interface elements, etc...). The Asset has, for now, an
Image, an Animation and Movement sub-structure. Each one is, in fact, a
pointer to a structure, which allows to define to NULL if the object
does not need it.

THE CODE DOES NOT WORK WITH THIS COMMIT. It lacks a function to populate
the Asset structure with the images and other infos for each existing
object.