Add ways to save & generate the assets database
The assets sqlite3 is not in VCS. However, to be able to test the
software, it needs to be generated. Now there are two ways to regenerate
it:
  - the generate_db.py script creates the database and fills it, using
  the tsv_to_db.py script in the operation
  - a dump of the db is in VCS, and can be loaded.

A command to save the DB has been added to the Makefile, alongside to
shortcuts for loading the dump/using the script. The README file has
been updated to document their usage.